# web/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lucen AI",
    "description": "Synthetic Identity and Deepfake Insurance-Claim Fraud Detection System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analyze/image": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Image",
        "operationId": "analyze_image_api_v1_analyze_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_analyze_image_api_v1_analyze_image_post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analyze/document": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Document",
        "operationId": "analyze_document_api_v1_analyze_document_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_analyze_document_api_v1_analyze_document_post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analyze/claim": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Claim",
        "operationId": "analyze_claim_api_v1_analyze_claim_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_analyze_claim_api_v1_analyze_claim_post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/analytics/summary": {
      "get": {
        "tags": [
          "Analytics",
          "Analytics"
        ],
        "summary": "Get Summary",
        "description": "Returns dashboard executive overview KPIs:\n- claims_today, fast_tracked, flagged, flagged_rate\n- top_reasons, sparkline_7d, open_rings",
        "operationId": "get_summary_api_v1_analytics_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/trends": {
      "get": {
        "tags": [
          "Analytics",
          "Analytics"
        ],
        "summary": "Get Trends",
        "description": "Returns longitudinal time-series analytics, risk distributions,\nmodality breakdowns, recycled evidence spikes, and decision velocity.",
        "operationId": "get_trends_api_v1_analytics_trends_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(7d|30d|90d)$",
              "default": "30d",
              "title": "Range"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(motor|health|property)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/business": {
      "get": {
        "tags": [
          "Analytics",
          "Analytics"
        ],
        "summary": "Get Business Analytics",
        "description": "Computes business-focused metrics from REAL claims only.\nReturns: {\n    \"claims_processed\": int,\n    \"fast_track_rate\": float (%),\n    \"fraud_prevented_amount\": float,\n    \"amount_at_risk\": float,\n    \"roi_multiple\": float,\n    \"patterns\": list of dict,\n    \"assumptions\": dict\n}",
        "operationId": "get_business_analytics_api_v1_analytics_business_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(7d|30d|90d|all)$",
              "default": "30d",
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_analyze_claim_api_v1_analyze_claim_post": {
        "properties": {
          "image": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Image",
            "default": []
          },
          "images": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Images",
            "default": []
          },
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files",
            "default": []
          },
          "document": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          },
          "id_photo": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Photo"
          },
          "selfie": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selfie"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "Body_analyze_claim_api_v1_analyze_claim_post"
      },
      "Body_analyze_document_api_v1_analyze_document_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "document": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          }
        },
        "type": "object",
        "title": "Body_analyze_document_api_v1_analyze_document_post"
      },
      "Body_analyze_image_api_v1_analyze_image_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "image": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          }
        },
        "type": "object",
        "title": "Body_analyze_image_api_v1_analyze_image_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}